Fixed chdir in bindmount locations and symlinks
If the locations being deleted contains symlinks
to directories, then S_ISDIR was attempting rmdir
which fails.  Turns out remove() already does the
proper checks, so cleaned up the code.

Chdir was added to fix the issue whereby if you
launched a job while doing --chdir or existing in
one of the bindmount locations, you would actually
be loaded into the pre-bindmounted location until
you did a chdir.
This could cause the node memory to be filled and
not cleaned up.