Set XDG_RUNTIME_DIR the same as SLURM_TMPDIR
Slurm jobs presently inherit XDG_RUNTIME_DIR from the login node,
set to /run/user/<uid>, which Slurm doesn't manage, so it typically
does not exist on the compute nodes. We can however set it the same way
as SLURM_TMPDIR, so that applications that use XDG_RUNTIME_DIR do not
complain that that directory does not exist.